# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar Trust Escrow API
  version: 0.1.0
  description: |
    REST API for the Stellar Trust Escrow platform — a trust-minimised escrow
    service built on the Stellar blockchain.

    ## Authentication
    Most endpoints require a Bearer JWT token in the `Authorization` header.
    Admin endpoints additionally require the `x-admin-api-key` header.

    ## Pagination
    All collection endpoints return a standardised envelope:
    ```
    {
      "data": [...],
      "page": 1, "limit": 20, "total": 42,
      "totalPages": 3, "hasNextPage": true, "hasPreviousPage": false
    }
    ```

    ## Environments
    - **Testnet**: `https://api.test.stellar-trust-escrow.example.com`
    - **Mainnet**: `https://api.stellar-trust-escrow.example.com`
  contact:
    name: Stellar Trust Escrow
    url: https://stellar-trust-escrow.example.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  x-repository: https://github.com/Stellar-Trust-Escrow/stellar-trust-escrow
  x-docs-url: https://docs.stellar-trust-escrow.example.com
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.test.stellar-trust-escrow.example.com
    description: Testnet
  - url: https://api.stellar-trust-escrow.example.com
    description: Mainnet
...
security:
  - BearerAuth: []
paths:
...
  /api/events:
    get:
      tags: [Events]
      summary: List all indexed on-chain events
      operationId: listEvents
      security: []
      parameters:
        - name: eventType
          in: query
          schema:
            type: string
          example: escrow_created
        - name: escrowId
          in: query
          schema:
            type: string
        - name: fromLedger
          in: query
          schema:
            type: integer
          description: Inclusive lower bound ledger sequence
        - name: toLedger
          in: query
          schema:
            type: integer
          description: Inclusive upper bound ledger sequence
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEvents'
...
  /api/events/escrow/{escrowId}:
    get:
      tags: [Events]
      summary: List all indexed events for a specific escrow (cursor-paginated)
      operationId: listEscrowEvents
      security: []
      parameters:
        - name: escrowId
          in: path
          required: true
          schema:
            type: string
        - name: eventType
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Cursor-paginated escrow events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorPaginatedEvents'
  /api/events/{id}:
    get:
      tags: [Events]
      summary: Get a single indexed event by database ID
      operationId: getEvent
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Event detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexedEvent'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login
    AdminAuth:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Admin API key for protected admin endpoints
  parameters:
    Page:
      name: page
      in: query
      description: 1-based page number. Values below 1 are normalised to 1.
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      name: limit
      in: query
      description: Page size. Defaults to 20 and is capped at 100.
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor (an event id) from a previous response's `nextCursor`. Omit to fetch the first page.
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong
        errorId:
          type: string
          description: Sentry event ID for support correlation
    PaginatedResponse:
      type: object
      required: [data, page, limit, total, totalPages, hasNextPage, hasPreviousPage]
      properties:
        data:
          type: array
          items: {}
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3
        hasNextPage:
          type: boolean
          example: true
        hasPreviousPage:
          type: boolean
          example: false
...
    IndexedEvent:
      type: object
      properties:
        id:
          type: integer
          format: int64
        ledger:
          type: integer
        txHash:
          type: string
        eventType:
          type: string
        escrowId:
          type: string
          nullable: true
        data:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
    PaginatedEvents:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/IndexedEvent'
    CursorPaginatedEvents:
      type: object
      required: [events, nextCursor]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/IndexedEvent'
        nextCursor:
          type: string
          nullable: true
          description: Pass this as `cursor` to fetch the next page. `null` when there are no more events.
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...